Implemented delete, localStorage, and edit functionality

+++ scripts/app.js
@@ -1,6 +1,7 @@
 import handleUserName from "./header.js";
 import * as expenseModule from "./manage-expenses.js";
 
+
 handleUserName();
 
 
@@ -13,13 +14,44 @@ document.querySelectorAll('.js-expense-input')
         })
     });
 
-// fix the delete buttns
-document.querySelectorAll('.js-expense-delete')
-    .forEach(deleteBtn => {
-        deleteBtn.addEventListener('click', () => {
-            const { expenseId } = deleteBtn.dataset;
 
+// Event listeners for generated HTML
+document.addEventListener('DOMContentLoaded', () => {
+    document.querySelector('.js-expenses-grid').addEventListener('click', event => {
+        if (event.target.classList.contains('js-expense-delete') || event.target.closest('.js-expense-delete')) {
+            const { expenseId } = event.target.closest('.js-expense-delete').dataset; 
+        
             expenseModule.deleteExpense(expenseId);
-            console.log('X clicked'); 
-        });
+        } else if (event.target.classList.contains('js-edit-expense') || event.target.closest('.js-edit-expense')) {
+            const { expenseId } = event.target.closest('.js-edit-expense').dataset;
+            
+            if (!document.getElementById('edit-expense-input')) {
+                expenseModule.editExpense(expenseId);
+            } else {
+                console.error('Edit Input Box already created');
+                return;
+            }
+        } else if (event.target.classList.contains('js-save-edit-button') || event.target.closest('.js-edit-expense')) {
+            const { expenseId } = event.target.closest('.js-save-edit-button').dataset;
+            
+            expenseModule.saveNewExpense(expenseId);
+        }
+    });
+
+    document.querySelector('.js-expenses-grid').addEventListener('keydown', event => {
+        if (event.target.closest('#edit-expense-input')) {
+            if (event.key === 'Enter') {
+                const { expenseId } = event.target.closest('#edit-expense-input').dataset;
+                expenseModule.saveNewExpense(expenseId)
+            }
+
+        }
+    });
+
+});
+
+
+document.querySelector('.js-delete-all-expenses')
+    .addEventListener('click', () => {
+        expenseModule.deleteAllExpenses();
     });
